+++ CHANGELOG.md
@@ -18,6 +18,8 @@ Follows versioning (SemVers) of https://docs.unity3d.com/Manual/upm-semver.html
 ### Changes
 - Updated documentation to reflect the new changes to the Editor Window
 - Updated some styling on editor window to be more readable
+- Added checking for package path, throwing an error message if package wasn't found
+- Removed some unecessary console logs
 
 ## 1.2.2
 ### Added